requirements.txt
# 분석 레이어 의존성

# 구조화된 로깅
structlog>=23.0.0

# 데이터 검증
pydantic>=2.0.0

# Redis
redis>=5.0.0

# Gemini API
google-genai>=1.0.0
